feat: change the all id to Long

+++ src/main/java/com/finalproject/document/management/entity/Department.java
@@ -18,7 +18,7 @@ public class Department {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @JsonBackReference
