+++ src/main/java/com/finalproject/document/management/repository/DocumentTypeRepository.java
@@ -3,5 +3,5 @@ package com.finalproject.document.management.repository;
 import com.finalproject.document.management.entity.DocumentType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentTypeRepository extends JpaRepository<DocumentType, Integer> {
+public interface DocumentTypeRepository extends JpaRepository<DocumentType, Long> {
 }
