+++ src/main/java/com/finalproject/document/management/service/DepartmentService.java
@@ -3,5 +3,5 @@ package com.finalproject.document.management.service;
 import com.finalproject.document.management.entity.Department;
 
 public interface DepartmentService {
-    Department findById(int id);
+    Department findById(Long id);
 }
