+++ src/main/java/com/finalproject/document/management/service/DocumentStatusService.java
@@ -3,5 +3,5 @@ package com.finalproject.document.management.service;
 import com.finalproject.document.management.entity.DocumentStatus;
 
 public interface DocumentStatusService {
-    DocumentStatus findByID(int id);
+    DocumentStatus findByID(Long id);
 }
