+++ src/main/java/com/finalproject/document/management/service/DocumentTypeService.java
@@ -3,5 +3,5 @@ package com.finalproject.document.management.service;
 import com.finalproject.document.management.entity.DocumentType;
 
 public interface DocumentTypeService {
-    DocumentType findById(int id);
+    DocumentType findById(Long id);
 }
